Add "hh" to sprintf "d" formats to prevent format-overflow errors
The toolchain shipped with ESP-IDF has g++ version 8.4.0
which emits overflow errors when it thinks there's a possibility
of overrunning the static buffer given to sprintf.  In practical
terms, there's no possibility of that but the compiler doesn't
know that.  Making the sprintf formats use "%02hhd" ("half-half"
or "byte"), for the uint8_t returns from the BCDtoDEC calls gives
the compiler enough hints as to the printable length of the
parameter to suppress the error.